Move clustered tests to ClusteredEventBusTestBase
